Implement the product message formatter
